Merge main to Readme-update

+++ src/main/java/com/example/voting/model/CandidateTotalVote.java
@@ -11,5 +11,5 @@ import lombok.Setter;
 @NoArgsConstructor
 public class CandidateTotalVote {
     private String candidateName;
-    private int totalVotes;
+    private Long totalVotes;
 }
